chore: rename AllTypes.ts to AllType.ts
chore: add AllType.PrimitiveValues

chore: add Any.Analysis to AllType.Analysis
