+++ src/types/Record.ts
@@ -1,5 +1,5 @@
 import { typeSym, valueSym } from '../utils'
-import { AllType } from './AllTypes'
+import { AllType } from './AllType'
 import { ValueType } from './types'
 import { undef, Undefined } from './Undefined'
 import { union, Union } from './Union'
