+++ src/types/strictCheck.ts
@@ -1,4 +1,4 @@
-import { AllType } from './AllTypes'
+import { AllType } from './AllType'
 
 export function strictCheck<T extends AllType>(type: T, subject: unknown): subject is T {
